+++ src/main/java/com/ethercamp/harmony/dto/WalletInfoDTO.java
@@ -3,6 +3,7 @@ package com.ethercamp.harmony.dto;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 public class WalletInfoDTO {
 
-    private final Long totalAmount;
+    private final BigInteger totalAmount;
 
     private final List<WalletAddressDTO> addresses = new ArrayList();
 }
